+++ Backend/src/types/common.types.ts
@@ -1,0 +1,7 @@
+export enum DATE_RANGE {
+    TODAY = 'today',
+    LAST_WEEK = 'last_week',
+    LAST_MONTH = 'last_month',
+    LAST_3_MONTHS = 'last_3_months',
+    ALL_TIME = 'all_time',
+}