+++ Frontend/src/shared/components/sidebar/app-sidebar.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { CircleUser, Inbox, LucideIcon, Settings } from 'lucide-react';
+import { CircleUser, Folder, Inbox, LucideIcon, Mail, Settings } from 'lucide-react';
 import React, { useEffect, useState } from 'react';
 
 import { useGetAccountsQuery } from '@/modules/accounts/services/useAccountApi';
@@ -47,6 +47,18 @@ const data: SidebarData = {
                 {
                     title: 'All Mail',
                     url: HOME_ROUTES.UNIFIED_INBOX,
+                },
+            ],
+        },
+        {
+            title: 'Folders',
+            url: '#',
+            icon: Folder,
+            isActive: true,
+            items: [
+                {
+                    title: 'Account',
+                    url: '#',
                 },
             ],
         },
@@ -148,9 +160,25 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
                     url: HOME_ROUTES.ACCOUNT_INBOX(acc._id),
                 })),
             ];
+            const folderItems = [
+                { title: 'All Folders', url: HOME_ROUTES.ALL_FOLDERS },
+                ...accounts.map((acc) => ({
+                    title: acc.emailAddress,
+                    url: HOME_ROUTES.ACCOUNT_FOLDERS(acc._id),
+                })),
+            ];
+            const updatedNavMain = sidebarData.navMain.map((item) => {
+                if (item.title === 'Inbox') {
+                    return { ...item, items: inboxItems };
+                }
+                if (item.title === 'Folders') {
+                    return { ...item, items: folderItems };
+                }
+                return item;
+            });
             setSidebarData((prev) => ({
                 ...prev,
-                navMain: prev.navMain.map((item) => (item.title === 'Inbox' ? { ...item, items: inboxItems } : item)),
+                navMain: updatedNavMain,
             }));
         }
     }, [accounts]);